runner/requirements.txt
fastapi
uvicorn[standard]
pyyaml
clickhouse-connect
dbt-core
dbt-clickhouse

